Move *Object functions back to signer and verifier
Also removed extraneous testing
